feat: cache all markdown processing (excerptHtml, excerptPlainText, contentPlainText)
- Move excerpt HTML generation into reviewsMarkdown.ts
- Add contentPlainText caching to both reviews and pages
- Add excerptPlainText caching for reviews
- Remove redundant getContentPlainText functions
- Update schema version to v4 to bust existing cache

This eliminates all runtime markdown-to-HTML and markdown-to-plaintext
processing, moving it to the initial parse/cache phase. Should provide
significant build time improvements for individual review pages and
pages with excerpts.
