+++ src/api/data/pagesMarkdown.ts
@@ -1,7 +1,13 @@
 import matter from "gray-matter";
 import { promises as fs } from "node:fs";
 import path from "node:path";
+import { remark } from "remark";
+import remarkGfm from "remark-gfm";
+import smartypants from "remark-smartypants";
+import strip from "strip-markdown";
 import { z } from "zod";
+
+import { removeFootnotes } from "~/api/utils/markdown/removeFootnotes";
 
 import { ContentCache, generateSchemaHash } from "./utils/cache";
 import { getContentPath } from "./utils/getContentPath";
@@ -9,6 +15,7 @@ import { getContentPath } from "./utils/getContentPath";
 const pagesMarkdownDirectory = getContentPath("pages");
 
 type MarkdownPage = {
+  contentPlainText: string;
   rawContent: string;
   slug: string;
   title: string;
@@ -28,8 +35,9 @@ export async function allPagesMarkdown(): Promise<MarkdownPage[]> {
 
 async function getCache(): Promise<ContentCache<MarkdownPage>> {
   if (!cacheInstance) {
+    // v2: Added contentPlainText field to cached data
     const schemaHash = await generateSchemaHash(
-      JSON.stringify(DataSchema.shape),
+      JSON.stringify(DataSchema.shape) + "-v2-plain-text",
     );
     cacheInstance = new ContentCache<MarkdownPage>(
       "pages-markdown",
@@ -37,6 +45,10 @@ async function getCache(): Promise<ContentCache<MarkdownPage>> {
     );
   }
   return cacheInstance;
+}
+
+function getMastProcessor() {
+  return remark().use(remarkGfm).use(smartypants);
 }
 
 async function parseAllPagesMarkdown() {
@@ -57,7 +69,15 @@ async function parseAllPagesMarkdown() {
           const { content: rawContent, data } = matter(content);
           const greyMatter = DataSchema.parse(data);
 
+          // Generate plain text content
+          const contentPlainText = getMastProcessor()
+            .use(removeFootnotes)
+            .use(strip)
+            .processSync(rawContent)
+            .toString();
+
           return {
+            contentPlainText,
             rawContent,
             slug: greyMatter.slug,
             title: greyMatter.title,
