+++ src/pages/reviews/[slug]/_index.spec.ts
@@ -6,7 +6,7 @@ import { loadRenderers } from "astro:container";
 import * as prettier from "prettier";
 import { describe, it } from "vitest";
 
-import { allReviews, getContentPlainText } from "~/api/reviews";
+import { allReviews } from "~/api/reviews";
 import { normalizeScriptSrc } from "~/utils/normalizeScriptSrc";
 
 import Review from "./index.astro";
@@ -37,7 +37,7 @@ describe("/reviews/:slug", () => {
         {
           partial: false,
           props: {
-            contentPlainText: getContentPlainText(review.rawContent),
+            contentPlainText: review.contentPlainText,
             slug: review.slug,
           },
           request: new Request(
