+++ src/pages/reviews/[slug]/index.astro
@@ -1,5 +1,5 @@
 ---
-import { allReviews, getContentPlainText } from "~/api/reviews";
+import { allReviews } from "~/api/reviews";
 import { getProps } from "~/components/Review/getProps";
 import { Review } from "~/components/Review/Review";
 import Layout from "~/layouts/Layout.astro";
@@ -19,7 +19,7 @@ export async function getStaticPaths() {
           slug: review.slug,
         },
         props: {
-          contentPlainText: getContentPlainText(review.rawContent),
+          contentPlainText: review.contentPlainText,
           slug: review.slug,
         },
       };
